worker3/requirements.txt
redis
celery
Flask-SQLAlchemy
SQLAlchemy
marshmallow-sqlalchemy
psycopg2
moviepy
opencv-python
google-cloud-storage
google-cloud-pubsub
oauth2client
httplib2